refactor: 폴더 구조 개선
- 관심사별로 폴더를 만들어서 정리
