+++ src/components/LottoGame.js
@@ -1,7 +1,7 @@
 import PurchaseForm from "./PurchaseForm.js";
-import LottoPurchaseHistory from "./LottoPurchaseHistory.js";
-import LottoWinningInfoForm from "./LottoWinningInfoForm.js";
-import WinningStatistic from "./WinningStatistic.js";
+import LottoPurchaseHistory from "./lottoHistoryItem/LottoPurchaseHistory.js";
+import LottoWinningInfoForm from "./winningLottoInfoForm/LottoWinningInfoForm.js";
+import WinningStatistic from "./winningStatistic/WinningStatistic.js";
 import Modal from "./common/Modal.js";
 import LottoMachine from "../domain/LottoMachine.js";
 import LottoResult from "../domain/LottoResult.js";
